+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 1.0.9
+
+- Fixed minibuffer history order
+- Fixed keyboard navigation in the goto pane: focus no more flashes.
+
 ## 1.0.8
 
 - Added an 'undo panel' that shows actions history and supports right-clicking to go to any undo point.